feat(auth): expiration / max time to live as configuration
