+++ src/declarations/console/console.factory.did.mjs
@@ -39,7 +39,8 @@ export const idlFactory = ({ IDL }) => {
 		proposal_ids: IDL.Vec(IDL.Nat)
 	});
 	const AuthenticationConfigOpenIdDelegation = IDL.Record({
-		targets: IDL.Opt(IDL.Vec(IDL.Principal))
+		targets: IDL.Opt(IDL.Vec(IDL.Principal)),
+		max_time_to_live: IDL.Opt(IDL.Nat64)
 	});
 	const OpenIdProvider = IDL.Variant({ Google: IDL.Null });
 	const OpenIdProviderConfig = IDL.Record({ client_id: IDL.Text });
