+++ src/libs/auth/src/delegation/get.rs
@@ -1,8 +1,8 @@
+use crate::delegation::duration::build_expiration;
 use crate::delegation::seed::calculate_seed;
 use crate::delegation::targets::{build_targets, targets_to_bytes};
 use crate::delegation::types::{
-    Delegation, GetDelegationError, GetDelegationResult, OpenIdGetDelegationArgs, SessionKey,
-    SignedDelegation, Timestamp,
+    Delegation, GetDelegationError, GetDelegationResult, SessionKey, SignedDelegation,
 };
 use crate::openid::types::interface::{OpenIdCredential, OpenIdCredentialKey};
 use crate::state::get_salt;
@@ -14,32 +14,32 @@ use ic_canister_sig_creation::{delegation_signature_msg, DELEGATION_SIG_DOMAIN};
 use serde_bytes::ByteBuf;
 
 pub fn openid_get_delegation(
-    args: &OpenIdGetDelegationArgs,
+    session_key: &SessionKey,
     client_id: &OpenIdProviderClientId,
     credential: &OpenIdCredential,
     auth_heap: &impl AuthHeapStrategy,
     certificate: &impl AuthCertificateStrategy,
 ) -> GetDelegationResult {
     get_delegation(
+        session_key,
         &client_id,
         &OpenIdCredentialKey::from(credential),
-        &args.session_key,
-        &args.expiration,
         auth_heap,
         certificate,
     )
 }
 
 pub fn get_delegation(
+    session_key: &SessionKey,
     client_id: &str,
     key: &OpenIdCredentialKey,
-    session_key: &SessionKey,
-    expiration: &Timestamp,
     auth_heap: &impl AuthHeapStrategy,
     certificate: &impl AuthCertificateStrategy,
 ) -> GetDelegationResult {
     let seed = calculate_seed(client_id, key, &get_salt(auth_heap))
         .map_err(GetDelegationError::DeriveSeedFailed)?;
+
+    let expiration = build_expiration(auth_heap);
 
     let targets = build_targets(auth_heap);
 
@@ -49,7 +49,7 @@ pub fn get_delegation(
             seed: &seed,
             message: &delegation_signature_msg(
                 session_key,
-                *expiration,
+                expiration,
                 targets_to_bytes(&targets).as_ref(),
             ),
         };
@@ -64,7 +64,7 @@ pub fn get_delegation(
             Ok(signature) => Ok(SignedDelegation {
                 delegation: Delegation {
                     pubkey: session_key.clone(),
-                    expiration: *expiration,
+                    expiration,
                     targets,
                 },
                 signature: ByteBuf::from(signature),
