+++ src/libs/auth/src/delegation/mod.rs
@@ -1,4 +1,5 @@
 mod constants;
+mod duration;
 mod get;
 mod impls;
 mod prepare;
