+++ src/libs/satellite/src/auth/delegation.rs
@@ -32,7 +32,7 @@ pub async fn openid_prepare_delegation(
     // TODO: create and assert user
 
     let result = delegation::openid_prepare_delegation(
-        args,
+        &args.session_key,
         &client_id,
         &credential,
         &AuthHeap,
@@ -61,7 +61,7 @@ pub fn openid_get_delegation(
     };
 
     let result = delegation::openid_get_delegation(
-        args,
+        &args.session_key,
         &client_id,
         &credential,
         &AuthHeap,
