+++ src/tests/declarations/test_satellite/test_satellite.factory.did.js
@@ -9,8 +9,7 @@ export const idlFactory = ({ IDL }) => {
 	const OpenIdPrepareDelegationArgs = IDL.Record({
 		jwt: IDL.Text,
 		session_key: IDL.Vec(IDL.Nat8),
-		salt: IDL.Vec(IDL.Nat8),
-		max_time_to_live: IDL.Opt(IDL.Nat64)
+		salt: IDL.Vec(IDL.Nat8)
 	});
 	const AuthenticateUserArgs = IDL.Variant({
 		OpenId: OpenIdPrepareDelegationArgs
@@ -136,7 +135,8 @@ export const idlFactory = ({ IDL }) => {
 		version: IDL.Opt(IDL.Nat64)
 	});
 	const AuthenticationConfigOpenIdDelegation = IDL.Record({
-		targets: IDL.Opt(IDL.Vec(IDL.Principal))
+		targets: IDL.Opt(IDL.Vec(IDL.Principal)),
+		max_time_to_live: IDL.Opt(IDL.Nat64)
 	});
 	const OpenIdProvider = IDL.Variant({ Google: IDL.Null });
 	const OpenIdProviderConfig = IDL.Record({ client_id: IDL.Text });
